+++ src/main/java/com/tune/reporting/helpers/ReportExportWorker.java
@@ -4,7 +4,7 @@ package com.tune.reporting.helpers;
  * ReportExportWorker.java
  *
  * <p>
- * Copyright (c) 2014 Tune, Inc
+ * Copyright (c) 2014 TUNE, Inc.
  * All rights reserved.
  * </p>
  *
@@ -38,9 +38,9 @@ package com.tune.reporting.helpers;
  * @category  tune-reporting
  * @package   com.tune.reporting
  * @author    Jeff Tanner jefft@tune.com
- * @copyright 2014 Tune (http://www.tune.com)
+ * @copyright 2014 TUNE, Inc. (http://www.tune.com)
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   $Date: 2014-12-12 05:24:55 $
+ * @version   $Date: 2014-12-24 13:23:15 $
  * @link      https://developers.mobileapptracking.com @endlink
  * </p>
  */
@@ -71,7 +71,7 @@ public class ReportExportWorker {
    */
   private String exportAction = null;
   /**
-   * User's Tune MobileAppTracking API Key.
+   * User's TUNE MobileAppTracking API Key.
    * @var String
    */
   private String apiKey = null;